+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.3.2](https://github.com/UVerify-io/uverify-ui/compare/v1.3.1...v1.3.2) (2025-04-24)
+
+
+### Bug Fixes
+
+* bump connect wallet version to fix eternl url ([970f17e](https://github.com/UVerify-io/uverify-ui/commit/970f17e14a5ed6d665a4f3bb865c2cfa3deb3e5f))
+
+## [1.3.1](https://github.com/UVerify-io/uverify-ui/compare/v1.3.0...v1.3.1) (2025-04-24)
+
+
+### Bug Fixes
+
+* correct wrong link to play and app store ([89d686d](https://github.com/UVerify-io/uverify-ui/commit/89d686d3deb651caa6c30937bb9559fa5fff90ba))
+
+## [1.3.0](https://github.com/UVerify-io/uverify-ui/compare/v1.2.1...v1.3.0) (2025-04-24)
+
+
+### Features
+
+* add mobile instructions and support for in-app browsers ([8e12c94](https://github.com/UVerify-io/uverify-ui/commit/8e12c9409d326243351cf60475d305b8ac6e072c))
+
 ## [1.2.1](https://github.com/UVerify-io/uverify-ui/compare/v1.2.0...v1.2.1) (2025-04-23)